refactor: restructure cli
