+++ src/command/generator.ts
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
-/** Generate utils and compare with current */
-import * as fs from 'fs';
+// Local
 import { getFilePaths, generateAST, getClassNames, filterClasses, writeCSS, readConfigFile, packageVersion } from '@/utils';
 
 /** run utils welcome - current version */
@@ -28,19 +27,5 @@ const rawClasses: string[] = filePaths.reduce((acc: string[], path) => {
 // 2 Filter utility classes, ex. "flex d-f ml-20" => "d-f ml-20"
 const classes = filterClasses(rawClasses)
 
-let generatedCSS: string = '';
-
-// 3 Format utilityClass into .CSS
-classes.forEach(({ fullClass, classKey, classValue }) => {
-  generatedCSS += `.${fullClass} { ${classKey}: ${classValue}; }\n`;
-});
-
-const currentCSS = fs.readFileSync(writeTo, 'utf-8').trim();
-
-if (currentCSS.replace(/\s+/g, '') === generatedCSS.replace(/\s+/g, '')) {
-    console.log('Generated css utilities are up to date.')
-    process.exit(0)
-} else {
-    console.error('Generated css utilities are outdated.')
-    process.exit(1)
-}
+// 3 Translate to CSS & writeTo path
+writeCSS({classes, filePath: writeTo})
