+++ src/command/watch.ts
@@ -1,4 +1,5 @@
 import { spawn } from 'child_process';
+// Local
 import { readConfigFile } from '@/utils';
 
 const { exclude = [], readFrom = "./", extensions = "tsx,ts,js,jsx" } = readConfigFile()
