+++ src/utils/cli-unsafe.ts
@@ -1,6 +1,7 @@
+import path from 'path';
+import fs from 'fs';
+// Local
 import { pathTo } from "@/utils";
-import * as path from 'path';
-import * as fs from 'fs';
 
 // /** Deprecated  @deprecated*/
 // export const installDependencies = (dependencies: string[]) => {
