+++ src/utils/css.ts
@@ -1,8 +1,9 @@
 import path from 'path';
 import fs from 'fs';
-import * as parser from '@babel/parser';
+import parser from '@babel/parser';
 import traverse from "@babel/traverse";
-import { shortKeys, shortValues } from '@/dictionary';
+// Local
+import { shortKeys, shortValues } from '@/constants';
 
 interface utilityClass {
     fullClass: string, // The raw input class name, ex. 'm-10'
