+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './css';
 export * from './cli-unsafe';
 export * from './cli';
+export {default as ProcessRetriever} from './Process';
